projects/mstc_py/requirements.txt
streamlit
pandas
supabase
requests
google-generativeai
python-dotenv
pydantic
